# packages/aio-commerce-lib-app/generated/openapi.gen.json
{
  "info": {
    "title": "AIO Commerce SDK API",
    "version": "1.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/installation/validation": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "appData": {
                    "type": "object",
                    "properties": {
                      "consumerOrgId": {
                        "type": "string",
                        "minLength": 1
                      },
                      "orgName": {
                        "type": "string",
                        "minLength": 1
                      },
                      "projectId": {
                        "type": "string",
                        "minLength": 1
                      },
                      "projectName": {
                        "type": "string",
                        "minLength": 1
                      },
                      "projectTitle": {
                        "type": "string",
                        "minLength": 1
                      },
                      "workspaceId": {
                        "type": "string",
                        "minLength": 1
                      },
                      "workspaceName": {
                        "type": "string",
                        "minLength": 1
                      },
                      "workspaceTitle": {
                        "type": "string",
                        "minLength": 1
                      }
                    },
                    "required": [
                      "consumerOrgId",
                      "orgName",
                      "projectId",
                      "projectName",
                      "projectTitle",
                      "workspaceId",
                      "workspaceName",
                      "workspaceTitle"
                    ]
                  },
                  "commerceBaseUrl": {
                    "type": "string"
                  },
                  "commerceEnv": {
                    "type": "string"
                  },
                  "ioEventsUrl": {
                    "type": "string"
                  },
                  "ioEventsEnv": {
                    "type": "string"
                  }
                },
                "required": [
                  "appData",
                  "commerceBaseUrl",
                  "commerceEnv",
                  "ioEventsUrl",
                  "ioEventsEnv"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean"
                    },
                    "summary": {
                      "type": "object",
                      "properties": {
                        "totalIssues": {
                          "type": "number"
                        },
                        "errors": {
                          "type": "number"
                        },
                        "warnings": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "totalIssues",
                        "errors",
                        "warnings"
                      ]
                    },
                    "result": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "path": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "meta": {
                          "type": "object",
                          "properties": {
                            "label": {
                              "type": "string"
                            },
                            "description": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "label"
                          ]
                        },
                        "issues": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "code": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              },
                              "severity": {
                                "anyOf": [
                                  {
                                    "const": "error"
                                  },
                                  {
                                    "const": "warning"
                                  },
                                  {
                                    "const": "info"
                                  }
                                ]
                              },
                              "details": {
                                "type": "object",
                                "propertyNames": {
                                  "type": "string"
                                },
                                "additionalProperties": {}
                              }
                            },
                            "required": [
                              "code",
                              "message",
                              "severity"
                            ]
                          }
                        },
                        "children": {
                          "$ref": "#/components/schemas/Ref4"
                        }
                      },
                      "required": [
                        "name",
                        "path",
                        "meta",
                        "issues",
                        "children"
                      ]
                    }
                  },
                  "required": [
                    "valid",
                    "summary",
                    "result"
                  ]
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "405": {
            "description": "Method not allowed"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Ref4": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "path": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "meta": {
              "type": "object",
              "properties": {
                "label": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                }
              },
              "required": [
                "label"
              ]
            },
            "issues": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string"
                  },
                  "message": {
                    "type": "string"
                  },
                  "severity": {
                    "anyOf": [
                      {
                        "const": "error"
                      },
                      {
                        "const": "warning"
                      },
                      {
                        "const": "info"
                      }
                    ]
                  },
                  "details": {
                    "type": "object",
                    "propertyNames": {
                      "type": "string"
                    },
                    "additionalProperties": {}
                  }
                },
                "required": [
                  "code",
                  "message",
                  "severity"
                ]
              }
            },
            "children": {
              "$ref": "#/components/schemas/Ref4"
            }
          },
          "required": [
            "name",
            "path",
            "meta",
            "issues",
            "children"
          ]
        }
      },
...
    }
  }
}